requirements.txt
playwright==1.40.0
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0
openai==1.3.0
google-generativeai==0.3.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlalchemy==2.0.23
pytest==7.4.3
flask==3.0.0
streamlit==1.28.0
